Crate aoc_grid

Modules

prelude

Structs

GridBuf

Multi dimensional abstraction over a single dimensional collection

GridCursor

GridCursorCompact

GridCursorIter

GridCursorWalk

GridExtend

GridExtendCursors

GridExtendIter

GridKeysIter

GridOffset

GridPrint

GridStorageClone

Wrapping iterator which clones items.

GridTransformIter

Enums

GridPosDisplayType

Traits

GridPos

Trait for values to use as indices for a GridView.

GridPosDisplay

GridPosDisplayAuto

GridPosParse

GridStorage

Trait for backing stores for a GridBuf

GridStorageIntoIter

Extra trait for GridStorage to support iteration.

GridStorageMut

Additional trait for backing stores which which can provide references to items

GridView

GridViewExtend

GridViewIter

GridViewPrint

Type Definitions

GridIter