Trait aoc_grid::GridStorageIntoIter

pub trait GridStorageIntoIter {
    type Item;
    type Iter: Iterator<Item = Self::Item>;

    // Required method
    fn storage_iter(&self) -> Self::Iter;
}

Extra trait for GridStorage to support iteration.

This is a separate trait to make the lifetimes work. It should be implemented on a reference to the storage, rather than directly. This allows us to capture the lifetime without polluting the main trait.

Required Associated Types

type Item

type Iter: Iterator<Item = Self::Item>

Required Methods

fn storage_iter(&self) -> Self::Iter